refactor: centralize the kent.sh docs URL and drop boilerplate comments
Extract the kent.sh origin + server-setup docs URL into a shared appLinks module
instead of a feature-local constant, and remove self-evident comments from the
Cargo release profile.
