+++ apps/desktop/src/appLinks.ts
@@ -1,0 +1,3 @@
+export const kentWebsiteOrigin = "https://kent.sh";
+
+export const serverSetupDocsUrl = `${kentWebsiteOrigin}/server#background-service`;
