--- apps/desktop/src/features/startup/serverSetup.ts
@@ -1,5 +1,0 @@
-// Canonical deep link into the public docs covering the background Kent server the
-// desktop client connects to over loopback. Install commands differ by platform and
-// install method, so the notice links to the docs (the source of truth) rather than
-// embedding commands that can drift out of date.
-export const serverSetupDocsUrl = "https://kent.sh/server#background-service";
